Refactor graph dialog into its own component
